+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `stepToRem` helper.
+
 ## [2.4.5] - 2019-05-09
 
 Fix: